Cleaning Up
- Increased the warning level and fixed new compiler warnings.
- Now warnings are treated as errors.
- Cleaned up the properties of the projects, moved some to common locations.
- The LineFormatter module is now just Formatter.

+++ ntUPSd.Core/ntUPSd.Formatter.cpp
@@ -17,7 +17,7 @@ DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE,
 OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 */
 #include "stdafx.h"
-#include "ntUPSd.LineFormatter.h"
+#include "ntUPSd.Formatter.h"
 
 namespace CTL
 {
@@ -34,7 +34,7 @@ namespace CTL
 				{
 				case '%':
 					// Escaping a percent (_%_), take the current format line and pass it to a recall.
-					rstrResult.Append(rpszFormat, pchPos - rpszFormat);
+					rstrResult.Append(rpszFormat, static_cast<int>(pchPos - rpszFormat));
 					rpszFormat = 1 + pchPos;
 					return TextImpl(rstrResult, rpszFormat);
 
@@ -75,7 +75,7 @@ namespace CTL
 
 	HRESULT Format::ToString(::ATL::CStringA &rstrResult, LPCSTR pszValue) noexcept
 	{
-		return ToString(rstrResult, pszValue, strlen(pszValue));
+		return ToString(rstrResult, pszValue, static_cast<int>(strlen(pszValue)));
 	}
 
 	HRESULT Format::ToString(::ATL::CStringA &rstrResult, LPCSTR pszValue, int cchValue) noexcept
